refactor: 복합 쿼리의 Repository를 각각 리턴하는 저장소로 통합
