+++ src/main/java/roomescape/domain/ReservationTimeRepository.java
@@ -1,8 +1,10 @@
 package roomescape.domain;
 
+import java.time.LocalDate;
 import java.time.LocalTime;
 import java.util.List;
 import java.util.Optional;
+import roomescape.domain.dto.AvailableTimeDto;
 
 public interface ReservationTimeRepository {
 
@@ -15,4 +17,6 @@ public interface ReservationTimeRepository {
     void deleteById(long id);
 
     boolean existsByStartAt(LocalTime time);
+
+    List<AvailableTimeDto> findAvailableReservationTimes(LocalDate date, long themeId);
 }
