+++ src/main/java/roomescape/domain/ThemeRepository.java
@@ -1,5 +1,6 @@
 package roomescape.domain;
 
+import java.time.LocalDate;
 import java.util.List;
 import java.util.Optional;
 
@@ -12,4 +13,6 @@ public interface ThemeRepository {
     Optional<Theme> findById(long id);
 
     void deleteById(long id);
+
+    List<Theme> findPopularThemesDateBetween(LocalDate startDate, LocalDate endDate, int limit);
 }
