+++ src/main/java/roomescape/presentation/ReservationTimeController.java
@@ -1,6 +1,7 @@
 package roomescape.presentation;
 
 import java.net.URI;
+import java.time.LocalDate;
 import java.util.List;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
@@ -9,9 +10,11 @@ import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 import roomescape.application.ReservationTimeService;
 import roomescape.application.dto.request.ReservationTimeRequest;
+import roomescape.application.dto.response.AvailableTimeResponse;
 import roomescape.application.dto.response.ReservationTimeResponse;
 
 @RestController
@@ -40,5 +43,12 @@ public class ReservationTimeController {
     public ResponseEntity<Void> delete(@PathVariable long id) {
         reservationTimeService.deleteById(id);
         return ResponseEntity.noContent().build();
+    }
+
+    @GetMapping("/available")
+    public ResponseEntity<List<AvailableTimeResponse>> findAvailableTimes(@RequestParam LocalDate date,
+                                                                          @RequestParam long themeId) {
+        List<AvailableTimeResponse> responses = reservationTimeService.findAvailableTimes(date, themeId);
+        return ResponseEntity.ok(responses);
     }
 }
